Update Athlete to require tokens at creation
When an athlete is created we want them to have an access token and a
refresh token in the database.

This is so that we can be certain that each athlete that is created
has everything we need for communicating with the Strava API.

It will also make creating an athlete and their tokens easier when they
authorize our application with Strava.