Refactor DataCollection calls in the client
Refactor the calls to DataCollection to start the collection in the
background on successful authorization of an athlete.

The GearLive now listens for data collection broadcasts and updates its
state as it receives messages from the DataCollection.

This also prevents infinite loops from occuring whene there is an error
in the DataCollection.